finished off SFT control
Former-commit-id: b7e1a6ce3d69c59a820d2c388b8ae5f8f1af6627 [formerly 7b732d59a418569915b1e9fd71b73de7394137db]
Former-commit-id: 10eda2d85bb4e58b192de7d4a18237501e7bd4f7

+++ Interface/gui/run.cpp
@@ -55,19 +55,38 @@ void Altair::run()
         return;
     }
 
-    if (!targetDirObject.removeRecursively())
+    const QStringList& files = targetDirObject.entryList(QDir::Files | QDir::Dirs | QDir::NoDotAndDotDot);
+
+    if (! files.isEmpty())
     {
-        Warning0(QString("Supprimer le répertoire au lancement"),
-                 QString("Il n'a pas été possible de nettoyer le répertoire %1 au lancement de LHX.\nNettoyer le répertoire et relancer.").arg(QDir::toNativeSeparators(path)));
-        processFinished(exitCode::shouldLaunchRAltairAlone);
-        return;
+        if (files.first() != "Bulletins.csv" && files.first() != "Table.csv")
+        {
+            if (QMessageBox::Cancel
+                  == QMessageBox::warning(this, QString("Attention"),
+                                                tr("Vous allez supprimer le répertoire %1\npour regénérer les bases. ").arg(QDir::toNativeSeparators(path)),
+                                                QMessageBox::Ok|QMessageBox::Cancel))
+                {
+
+                    processFinished(exitCode::shouldLaunchRAltairAlone);
+                    return;
+                }
+        }
+
+        if (!targetDirObject.removeRecursively())
+        {
+            Warning0(QString("Supprimer le répertoire au lancement"),
+                     QString("Il n'a pas été possible de nettoyer le répertoire %1 au lancement de LHX.\nNettoyer le répertoire et relancer.").arg(QDir::toNativeSeparators(path)));
+            processFinished(exitCode::shouldLaunchRAltairAlone);
+            return;
+        }
+        else
+        if (targetDirObject.mkpath(path) == false)
+        {
+            Warning0(QString("Répertoire"), QString("Le répertoire de sortie %1 n'a pas pu être créé. Relancer après avoir réglé le problème.").arg(path));
+            return;
+        }
     }
-    else
-    if (targetDirObject.mkpath(path) == false)
-    {
-        Warning0(QString("Répertoire"), QString("Le répertoire de sortie %1 n'a pas pu être créé. Relancer après avoir réglé le problème.").arg(path));
-        return;
-    }
+
 #ifdef DEBUG
     outputTextEdit->append(PROCESSING_HTML_TAG + tr("Validation du répertoire de sortie ") + path);
 #endif
